+++ CHANGELOG.md
@@ -5,7 +5,8 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
-<!-- ## Unreleased -->
+## Unreleased
+* Modified the `bundle` property in project build options to support the subset of `polymer-bundler` options which can be serialized in a JSON file.
 <!-- Add new, unreleased changes here. -->
 
 ## [3.3.0] - 2017-06-09